test: split bounty API route coverage
